+++ iiif/error.py
@@ -14,7 +14,7 @@ from xml.etree.ElementTree import ElementTree, Element
 # Namespace used in XML error response
 I3F_NS = "http://library.stanford.edu/iiif/image-api/ns/"
 
-class I3fError:
+class IIIFError:
 
     def __init__(self,code=500,parameter='unknown',text=None):
         self.code=code