+++ iiif/manipulator_netpbm.py
@@ -9,12 +9,12 @@ import os.path
 import glob
 import subprocess
 
-from error import I3fError
-from request import I3fRequest
-from manipulator import I3fManipulator
+from error import IIIFError
+from request import IIIFRequest
+from manipulator import IIIFManipulator
 
-class I3fManipulatorNetpbm(I3fManipulator):
-    """Module to manipulate and image according to i3f rules
+class IIIFManipulatorNetpbm(IIIFManipulator):
+    """Module to manipulate and image according to iiif rules
 
     All exceptions are raise as Error objects which directly
     determine the HTTP response.
@@ -25,8 +25,8 @@ class I3fManipulatorNetpbm(I3fManipulator):
     pnmdir = None
 
     def __init__(self):
-        super(I3fManipulatorNetpbm, self).__init__()
-        self.complianceLevel="http://i3f.example.org/compliance/level/1"
+        super(IIIFManipulatorNetpbm, self).__init__()
+        self.complianceLevel="http://iiif.example.org/compliance/level/1"
         if (self.filecmd is None):
             self.find_binaries()
 
@@ -64,7 +64,7 @@ class I3fManipulatorNetpbm(I3fManipulator):
 
     def do_first(self):
         pid = os.getpid()
-        self.basename = os.path.join(self.tmpdir,'i3f_netpbm_'+str(pid))
+        self.basename = os.path.join(self.tmpdir,'iiif_netpbm_'+str(pid))
         outfile = self.basename+'.pnm'
         # Convert source file to pnm
         filetype = self.file_type(self.srcfile)
@@ -72,9 +72,9 @@ class I3fManipulatorNetpbm(I3fManipulator):
             #print "doing png->pnm"
             #print self.pngtopnm+' '+self.srcfile+' > '+outfile
             if (self.shell_call(self.pngtopnm+' '+self.srcfile+' > '+outfile)):
-                raise I3fError(text="Oops... got nonzero output from pngtopnm.")
+                raise IIIFError(text="Oops... got nonzero output from pngtopnm.")
         else:
-            raise I3fError(code='501',
+            raise IIIFError(code='501',
                            text='bad input file format')
         self.tmpfile = outfile
 
@@ -92,7 +92,7 @@ class I3fManipulatorNetpbm(I3fManipulator):
         else:
             #print "region: (%d,%d,%d,%d)" % (x,y,w,h)
             if (self.shell_call('cat '+infile+' | '+self.pnmcut+' '+str(x)+' '+str(y)+' '+str(w)+' '+str(h)+'  > '+outfile)):
-                raise I3fError(text="Oops... got nonzero output from pnmcut.")
+                raise IIIFError(text="Oops... got nonzero output from pnmcut.")
             self.width=w
             self.height=h
             self.tmpfile = outfile
@@ -109,7 +109,7 @@ class I3fManipulatorNetpbm(I3fManipulator):
         else:
             #print "size: scaling to (%d,%d)" % (w,h)
             if (self.shell_call('cat '+infile+' | '+self.pnmscale+' -width '+str(w)+' -height '+str(h)+'  > '+outfile)):
-                raise I3fError(text="Oops... got nonzero output from pnmscale.")
+                raise IIIFError(text="Oops... got nonzero output from pnmscale.")
             width=w
             height=h
             self.tmpfile=outfile
@@ -129,14 +129,14 @@ class I3fManipulatorNetpbm(I3fManipulator):
                 rot-=360.0
             #print "rotation: by %f degrees clockwise" % (rot)
             if (self.shell_call('cat '+infile+' | '+self.pnmrotate+' -background=#FFF '+str(-rot)+'  > '+outfile)):
-                raise I3fError(text="Oops... got nonzero output from pnmrotate.")
+                raise IIIFError(text="Oops... got nonzero output from pnmrotate.")
             self.tmpfile=outfile
         else:
             # Between 90 and 270 = flip and then -90 to 90
             rot-=180.0
             #print "rotation: by %f degrees clockwise" % (rot)
             if (self.shell_call('cat '+infile+' | '+self.pnmflip+' -rotate180 | '+self.pnmrotate+' '+str(-rot)+'  > '+outfile)):
-                raise I3fError(text="Oops... got nonzero output from pnmrotate.")
+                raise IIIFError(text="Oops... got nonzero output from pnmrotate.")
             self.tmpfile=outfile
 
     def do_quality(self):
@@ -147,18 +147,18 @@ class I3fManipulatorNetpbm(I3fManipulator):
         if (quality == 'grey'):
             #print "quality: grey"
             if (self.shell_call('cat '+infile+' | '+self.ppmtopgm+' > '+outfile)):
-                raise I3fError(text="Oops... got nonzero output from ppmtopgm.")
+                raise IIIFError(text="Oops... got nonzero output from ppmtopgm.")
             self.tmpfile=outfile
         elif (quality == 'bitonal'):
             #print "quality: grey"
             if (self.shell_call('cat '+infile+' | '+self.ppmtopgm+' | '+self.pamditherbw+' > '+outfile)):
-                raise I3fError(text="Oops... got nonzero output from ppmtopgm.")
+                raise IIIFError(text="Oops... got nonzero output from ppmtopgm.")
             self.tmpfile=outfile
         elif (quality == 'quality'):
             #print "quality: quality"
             self.tmpfile=infile
         else: 
-            raise I3fError(code=400,parameter='quality',
+            raise IIIFError(code=400,parameter='quality',
                            text="Unknown quality parameter value requested.")
 
     def do_format(self):
@@ -176,26 +176,26 @@ class I3fManipulatorNetpbm(I3fManipulator):
         if (fmt == 'png'):
             #print "format: png"
             if (self.shell_call(self.pnmtopng+' '+infile+' > '+outfile)):
-                raise I3fError(text="Oops... got nonzero output from pnmtopng.")
+                raise IIIFError(text="Oops... got nonzero output from pnmtopng.")
             mime_type="image/png"
         elif (fmt == 'tiff' or fmt == 'jp2'):
             #print "format: tiff/jp2"
             if (self.shell_call(self.pnmtotiff+' '+infile+' > '+outfile)):
-                raise I3fError(text="Oops... got nonzero output from pnmtotiff.")
+                raise IIIFError(text="Oops... got nonzero output from pnmtotiff.")
             mime_type="image/tiff"
             if (fmt == 'jp2'):
                 # use djatoka after tiff
                 if (self.shell_call(DJATOKA_COMP+' -i '+outfile+' -o '+outfile_jp2)):
-                    raise I3fError(text="Oops... got nonzero output from DJATOKA_COMP.")
+                    raise IIIFError(text="Oops... got nonzero output from DJATOKA_COMP.")
                 mime_type="image/jp2"
                 outfile=tmpfile_jp2
         elif (fmt == 'jpg'):
             #print "format: jpg"
             if (self.shell_call(self.pnmtojpeg+' '+infile+' > '+outfile)):
-                raise I3fError(text="Oops... got nonzero output from pnmtojpeg.")
+                raise IIIFError(text="Oops... got nonzero output from pnmtojpeg.")
             mime_type="image/jpeg"
         else:
-            raise I3fError(code=415, parameter='format',
+            raise IIIFError(code=415, parameter='format',
                            text="Unsupported output file format (%s), only png,jpg,tiff are supported."%(fmt))
         self.outfile=outfile
         self.output_format=fmt
@@ -223,7 +223,7 @@ class I3fManipulatorNetpbm(I3fManipulator):
         pout.close()
         m=re.search(', (\d+) by (\d+) ',pnmfileout);
         if (m is None):
-            raise I3fError(text="Bad output from pnmfile when trying to get size.")
+            raise IIIFError(text="Bad output from pnmfile when trying to get size.")
         w=int(m.group(1))
         h=int(m.group(2))
         #print "pnmfile output = %s" % (pnmfileout)