+++ CHANGES.rst
@@ -1,4 +1,4 @@
-i3f changelog
+iiif changelog
 =============
 
 2012-03-21 v0.1.0
@@ -7,8 +7,8 @@ i3f changelog
 2013-05-21 v0.2.0
 - Reorganized, aims for IIIF API v1.0
   (http://www-sul.stanford.edu/iiif/image-api/)
-- PIL library support (I3fManipulatorPIL) and null tested, the netpbm
-  version (I3fManipulatorNetpbm) has not been updated/tested and is 
+- PIL library support (IIIFManipulatorPIL) and null tested, the netpbm
+  version (IIIFManipulatorNetpbm) has not been updated/tested and is 
   likely broken.
 - Designed to work with Loris for for Open Seadragon demo
   (https://github.com/zimeon/loris)